// public-site/src/lib/schema.ts
export type Review = {
  id: number;
  author: string;
  rating: number;
  text: string;
  date: string | null; // e.g. "2 weeks ago"
  isActive: boolean;
  createdAt: Date;
};